+++ src-tauri/src/feat/mod.rs
@@ -1,0 +1,14 @@
+mod backup;
+mod clash;
+mod config;
+mod profile;
+mod proxy;
+mod window;
+
+// Re-export all functions from modules
+pub use backup::*;
+pub use clash::*;
+pub use config::*;
+pub use profile::*;
+pub use proxy::*;
+pub use window::*;
